frontEnd/getCampaigns/requirements.txt
functions-framework==3.4.0
flask
google-cloud-bigquery
